requirements.txt
anyio==3.6.2
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
fastapi==0.95.0
h11==0.14.0
httptools==0.5.0
idna==3.4
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
sniffio==1.3.0
soupsieve==2.4
starlette==0.26.1
typing_extensions==4.5.0
unicode==2.9
Unidecode==1.3.6
urllib3==1.26.15
uvicorn==0.21.1
watchfiles==0.19.0
websockets==11.0
